requirements.txt
numpy
pandas
seaborn
matplotlib
xgboost
catboost
python-dotenv
pymysql
mysql-connector-python
scikit-learn
mlflow
dagshub
# -e .
gunicorn